# web/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AppConfig": {
        "additionalProperties": false,
        "description": "Complete configuration with reference integrity checks.",
        "properties": {
          "alert_targets": {
            "items": {
              "discriminator": {
                "mapping": {
                  "mqtt": "#/components/schemas/MqttTarget",
                  "script": "#/components/schemas/ScriptTarget",
                  "webhook": "#/components/schemas/WebhookTarget"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/MqttTarget"
                },
                {
                  "$ref": "#/components/schemas/WebhookTarget"
                },
                {
                  "$ref": "#/components/schemas/ScriptTarget"
                }
              ]
            },
            "maxItems": 128,
            "title": "Alert Targets",
            "type": "array"
          },
          "discovery": {
            "$ref": "#/components/schemas/DiscoveryConfig"
          },
          "sources": {
            "items": {
              "discriminator": {
                "mapping": {
                  "file": "#/components/schemas/FileSource",
                  "rtlsdr": "#/components/schemas/RtlSdrSource",
                  "soundcard": "#/components/schemas/SoundcardSource",
                  "stream": "#/components/schemas/StreamSource"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SoundcardSource"
                },
                {
                  "$ref": "#/components/schemas/StreamSource"
                },
                {
                  "$ref": "#/components/schemas/RtlSdrSource"
                },
                {
                  "$ref": "#/components/schemas/FileSource"
                }
              ]
            },
            "maxItems": 16,
            "title": "Sources",
            "type": "array"
          },
          "tone_sets": {
            "items": {
              "$ref": "#/components/schemas/ToneSet"
            },
            "maxItems": 500,
            "title": "Tone Sets",
            "type": "array"
          }
        },
        "title": "AppConfig",
        "type": "object"
      },
...
      "DiscoveryConfig": {
        "additionalProperties": false,
        "description": "Global tone auto-discovery settings.",
        "properties": {
          "clip": {
            "default": true,
            "title": "Clip",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "max_gap_s": {
            "default": 0.5,
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Max Gap S",
            "type": "number"
          },
          "max_segment_s": {
            "default": 3.0,
            "exclusiveMinimum": 0.0,
            "maximum": 10.0,
            "title": "Max Segment S",
            "type": "number"
          },
          "min_segment_s": {
            "default": 0.3,
            "exclusiveMinimum": 0.0,
            "maximum": 3.0,
            "title": "Min Segment S",
            "type": "number"
          },
          "tol_pct": {
            "default": 1.5,
            "maximum": 5.0,
            "minimum": 0.5,
            "title": "Tol Pct",
            "type": "number"
          }
        },
        "title": "DiscoveryConfig",
        "type": "object"
      },
      "FileSource": {
        "additionalProperties": false,
        "properties": {
          "discovery_enabled": {
            "default": true,
            "title": "Discovery Enabled",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "loop": {
            "default": false,
            "title": "Loop",
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "path": {
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "realtime": {
            "default": true,
            "title": "Realtime",
            "type": "boolean"
          },
          "tonesets": {
            "anyOf": [
              {
                "items": {
                  "minLength": 1,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "const": "all",
                "type": "string"
              }
            ],
            "default": "all",
            "title": "Tonesets"
          },
          "type": {
            "const": "file",
            "default": "file",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "path"
        ],
        "title": "FileSource",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MqttTarget": {
        "additionalProperties": false,
        "properties": {
          "broker": {
            "default": "localhost",
            "title": "Broker",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "events": {
            "items": {
              "enum": [
                "pre_alert",
                "recording_ready",
                "closed",
                "tone_discovered"
              ],
              "type": "string"
            },
            "title": "Events",
            "type": "array"
          },
          "ha_discovery": {
            "default": true,
            "title": "Ha Discovery",
            "type": "boolean"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "port": {
            "default": 1883,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "type": "integer"
          },
          "source": {
            "default": "manual",
            "enum": [
              "manual",
              "supervisor"
            ],
            "title": "Source",
            "type": "string"
          },
          "tls": {
            "default": false,
            "title": "Tls",
            "type": "boolean"
          },
          "topic": {
            "default": "tonewatch",
            "title": "Topic",
            "type": "string"
          },
          "type": {
            "const": "mqtt",
            "default": "mqtt",
            "title": "Type",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "MqttTarget",
        "type": "object"
      },
...
      "RtlSdrSource": {
        "additionalProperties": false,
        "properties": {
          "discovery_enabled": {
            "default": true,
            "title": "Discovery Enabled",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "freq_hz": {
            "title": "Freq Hz",
            "type": "number"
          },
          "gain": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gain"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "ppm": {
            "default": 0,
            "title": "Ppm",
            "type": "integer"
          },
          "squelch": {
            "default": 0,
            "title": "Squelch",
            "type": "integer"
          },
          "tonesets": {
            "anyOf": [
              {
                "items": {
                  "minLength": 1,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "const": "all",
                "type": "string"
              }
            ],
            "default": "all",
            "title": "Tonesets"
          },
          "type": {
            "const": "rtlsdr",
            "default": "rtlsdr",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "freq_hz"
        ],
        "title": "RtlSdrSource",
        "type": "object"
      },
      "ScriptTarget": {
        "additionalProperties": false,
        "properties": {
          "args": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Args",
            "type": "array"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "events": {
            "items": {
              "enum": [
                "pre_alert",
                "recording_ready",
                "closed",
                "tone_discovered"
              ],
              "type": "string"
            },
            "title": "Events",
            "type": "array"
          },
          "executable": {
            "minLength": 1,
            "title": "Executable",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "timeout_s": {
            "default": 30,
            "exclusiveMinimum": 0.0,
            "title": "Timeout S",
            "type": "number"
          },
          "type": {
            "const": "script",
            "default": "script",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "executable"
        ],
        "title": "ScriptTarget",
        "type": "object"
      },
...
      "ToneSet": {
        "additionalProperties": false,
        "description": "An ordered tone sequence and its alert policy.",
        "properties": {
          "alert_targets": {
            "default": [],
            "items": {
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              "type": "string"
            },
            "title": "Alert Targets",
            "type": "array"
          },
          "cooldown_s": {
            "default": 60,
            "minimum": 0.0,
            "title": "Cooldown S",
            "type": "number"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "max_gap_s": {
            "default": 0.5,
            "exclusiveMinimum": 0.0,
            "title": "Max Gap S",
            "type": "number"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "record": {
            "$ref": "#/components/schemas/RecordingPolicy"
          },
          "sequence": {
            "items": {
              "$ref": "#/components/schemas/ToneSpec"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Sequence",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "sequence"
        ],
        "title": "ToneSet",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookTarget": {
        "additionalProperties": false,
        "properties": {
          "allow_insecure_http": {
            "default": false,
            "title": "Allow Insecure Http",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "events": {
            "items": {
              "enum": [
                "pre_alert",
                "recording_ready",
                "closed",
                "tone_discovered"
              ],
              "type": "string"
            },
            "title": "Events",
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "include_audio": {
            "default": false,
            "title": "Include Audio",
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "secret": {
            "default": "",
            "title": "Secret",
            "type": "string"
          },
          "type": {
            "const": "webhook",
            "default": "webhook",
            "title": "Type",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url"
        ],
        "title": "WebhookTarget",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ToneWatch API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/alert-targets": {
      "get": {
        "operationId": "list_items_api_alert_targets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "title": "Response List Items Api Alert Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Items",
        "tags": [
          "configuration"
        ]
      },
      "post": {
        "operationId": "create_item_api_alert_targets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Item Api Alert Targets Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Item",
        "tags": [
          "configuration"
        ]
      }
    },
...
    "/api/config": {
      "get": {
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "configuration"
        ]
      },
      "put": {
        "operationId": "replace_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Config",
        "tags": [
          "configuration"
        ]
      }
    },
...
    "/api/sources": {
      "get": {
        "operationId": "list_items_api_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "title": "Response List Items Api Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Items",
        "tags": [
          "configuration"
        ]
      },
      "post": {
        "operationId": "create_item_api_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Item Api Sources Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Item",
        "tags": [
          "configuration"
        ]
      }
    },
...
    "/api/tonesets": {
      "get": {
        "operationId": "tonesets_api_tonesets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "title": "Response Tonesets Api Tonesets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Tonesets",
        "tags": [
          "configuration"
        ]
      },
      "post": {
        "operationId": "create_toneset_api_tonesets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Toneset Api Tonesets Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Toneset",
        "tags": [
          "configuration"
        ]
      }
    },
    "/api/tonesets/{item_id}": {
      "delete": {
        "operationId": "delete_toneset_api_tonesets__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Delete Toneset Api Tonesets  Item Id  Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Toneset",
        "tags": [
          "configuration"
        ]
      },
      "get": {
        "operationId": "get_toneset_api_tonesets__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Toneset Api Tonesets  Item Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Toneset",
        "tags": [
          "configuration"
        ]
      },
      "put": {
        "operationId": "update_toneset_api_tonesets__item_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToneSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Toneset Api Tonesets  Item Id  Put"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Toneset",
        "tags": [
          "configuration"
        ]
      }
    },
    "/api/tonesets/{item_id}/test": {
      "post": {
        "operationId": "test_toneset_api_tonesets__item_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Test Toneset Api Tonesets  Item Id  Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Toneset",
        "tags": [
          "configuration"
        ]
      }
    },
...
  }
}